Remove connection magic of Ohm.
This change, spawned by @djanowski, relies on the redis
client, in this case Redic, to be thread safe.

Instead of having a #connect method, we simply store
a `@redis` class variable in Ohm, and all models
inherits from that by default.

For cases where certain models need to connect to different
databases, they simple have to override that, i.e.

    User.redis = Redic.new(ENV["REDIS_URL2"])