Centralize Lua logic into a module.
This is similar to the first iteration we did a year ago,
but cheats a bit by pre-caching the SHA1s of each individual
lua file.

In addition, we use SCRIPT EXISTS - SCRIPT LOAD mechanics
as opposed to an EVAL FAIL EVALSHA1 method.

SCRIPT EXISTS results are memoized in-ruby (on a per
redis instance - lua file pair) to avoid unnecessary
repeated lookups.